+++ docs/CHANGELOG.md
@@ -36,3 +36,4 @@
 - Added encoding_scheme in catalog/algorithm.
 - Added deployment_source in catalog/agent.
 - Added agent_instance_ids and algorithm_instance_ids in pipeline.
+- Added endpoint for LCP communication.